Classe Pai DAO implementada, Antiga Classe Dao agora é a classe filha DAOTexto e a lista de dados se tornou uma lista estática
