+++ .gitignore
@@ -1,2 +1,7 @@
 /CRUD.dao/nbproject/private/
 /CRUD.Shared/nbproject/private/
+/CRUD.Business/dist/
+/CRUD.Business/build/
+/CRUD.Shared/build/
+/CRUD.dao/build/
+/CRUD.Shared/dist/
